PasteAs - command for pasting already copied files (from clipboard) under a different name
Pastes the file under a new name, with the original extension or a new one.

The "keep original extension" decides whether the original extension will be put as the last one,
so copying "test.txt" as "test.bla" with it checked will result in "test.bla.txt", so to paste as
"test.txt.bak" you will need to uncheck it.

In case overwriting should happen - a dialog will be displayed.

OLD notes (all done):

Still need to fix so that if you enter just a name - it pastes with old extension,
but if you enter this.is.a.test is pastes it like that.

Ideally, also include a button to allow pasting without an extension, which will toggle
the whole "if no extension - add old, otherwise keep as is" new filename fixing.

The "paste as..." command, useful for pasting any file under a customizable name.

EDIT: reworked the whole "old extension" / "new extension" by not scanning the input
text for extensions but rather using a checkbox to determine whether the file should
inherit the old extension or use the one from the input box

Just need to add a "overwrite" dialog logic in case a file like that already exists...